Isolate configuration application, export, and reconnect verification
Move connected configuration handling into a dedicated action runtime covering
--set, YAML configuration import, configuration export, direct owner/content/
position/channel URL updates, protobuf section traversal, and firmware settings
transactions.

Centralize channel URL stability checks and post-write reconnect verification so
transport loss, device reboot, configuration reload, and value verification are
handled by one coherent workflow. Preserve write pacing, destination handling,
normalization, and compatibility hooks that callers and tests already depend on.

Keep meshtastic.__main__ wrappers as stable adapters to the extracted helpers.
This makes configuration behavior independently testable while preserving CLI
syntax, output responsibilities, and existing import seams.
